+++ packages/app/src/app/backend/calls/addMetaAttrRemote.ts
@@ -5,9 +5,10 @@ const {url, params} = endpoints.addMetaAttrRemote;
 export const addMetaAttrRemote = ({
   clusterName,
   resourceId,
+  isStonith,
   name,
   value,
 }: {clusterName: string} & Parameters<typeof params>[0]): CallResult =>
   http.post(url({clusterName}), {
-    params: params({resourceId, name, value}),
+    params: params({resourceId, isStonith, name, value}),
   });